requirements.txt
azure-identity
azure-storage-file-datalake
pyodbc
pandas
pytest
faker
